feat: drop old arch and use clean arch for server and lib utilities
