+++ drizzle.config.ts
@@ -4,7 +4,7 @@ import { defineConfig } from 'drizzle-kit'
 config({ path: '.env' })
 
 export default defineConfig({
-  schema: './database/schema.ts',
+  schema: './server/schemas/index.ts',
   out: './supabase/migrations',
   dialect: 'postgresql',
   dbCredentials: { url: process.env.NEXT_DATABASE_URL }
