+++ server/actions/priorities.ts
@@ -1,6 +1,6 @@
 'use server'
 
-import { prioritiesMethods as priorities } from '@/database'
+import { prioritiesMethods as priorities } from '@/server/data/priorities'
 
 export async function getPriorities() {
   return await priorities.get()
