+++ server/actions/roles.ts
@@ -1,6 +1,6 @@
 'use server'
 
-import { rolesMethods as roles } from '@/database'
+import { rolesMethods as roles } from '@/server/data/roles'
 
 export async function getRoles() {
   return await roles.get()
