+++ server/actions/status.ts
@@ -1,6 +1,6 @@
 'use server'
 
-import { statusMethods as status } from '@/database'
+import { statusMethods as status } from '@/server/data/status'
 
 export async function getStatus() {
   return await status.get()
