+++ server/data/roles.ts
@@ -1,0 +1,7 @@
+import { database } from '@/server/config/database'
+
+export const rolesMethods = {
+  get: async () => {
+    return await database.query.roles.findMany()
+  }
+}
