+++ server/schemas/tickets.ts
@@ -1,10 +1,10 @@
 import { integer, timestamp, pgTable, text, serial } from 'drizzle-orm/pg-core'
 import { relations } from 'drizzle-orm'
 
-import { users } from '@/database/schemas/users'
-import { status } from '@/database/schemas/status'
-import { resolutions } from '@/database/schemas/resolutions'
-import { priorities } from '@/database/schemas/priorities'
+import { users } from '@/server/schemas/users'
+import { status } from '@/server/schemas/status'
+import { resolutions } from '@/server/schemas/resolutions'
+import { priorities } from '@/server/schemas/priorities'
 
 export const tickets = pgTable('tickets', {
   id: text('id')
